Added search functionality and null checks in UserController
In the UserController.cs file, a new `Search` method has been added that takes a string parameter and searches for it in the `userlist` based on the `Name` or `Email` properties of the users. If no results are found, it returns a `HttpNotFound` response. Otherwise, it returns a view named "Index" with the search results. The `Create` method has been updated to check if the `Name` or `Email` properties of the user are null before adding the user to the `userlist`. If either property is null, it returns the "Create" view with the user data. Otherwise, it adds the user to the `userlist`, assigns an ID to the user, and redirects to the "Index" view. The `Index.cshtml` file has been updated to include a form for the new `Search` method. The form includes a text box for the search term and a submit button. The table displaying the user data has been updated to be inside the form. The table now displays the `Name` and `Email` properties of each user in the `Model`, along with links to "Edit", "Details", and "Delete" actions for each user.